# docs/api/E2-identity/openapi.yaml
openapi: 3.1.0
info:
  title: hris-outsource · E2 Identity, Org & Master Data
  version: 1.0.0
  description: |
    OpenAPI 3.1 contract for **E2 — Identity, Org & Master Data**.

    > **NOTE — Profile change-requests removed 2026-06-14 (EPICS §8 E11):**
    > agent self-edits to phone, emergency contact, and bank account are now
    > **instant** (applied immediately, no approval). The `change-requests`
    > tag, paths, and schemas are deleted. Configurable approval now lives only
    > in E11 (leave/overtime).

    Scope: the people the system is about (employees + linked logins), the
    employment relationship (PKWT/PKWTT agreements + compensation), the client
    company directory, and the operational master data (leave types, attendance
    codes, overtime rules). Position is **free-text** (no master) with a
    typeahead search endpoint over existing values.

    Binding conventions: see `docs/api/CONVENTIONS.md` (sections referenced
    inline as §N). Picker endpoints are co-located with their owning resource
    per §18 — no separate `/lookups` namespace.

    Cross-references: employee-detail tab data (Penempatan/Kehadiran/Cuti/Lembur)
    is **not served** by E2; the E2 employee-detail response carries
    `x-cross-epic-references` mapping each tab to the owning epic's endpoint.

    Personas used in examples: **Sari Hadi (HR Admin)**, **Rudi Wijaya (Shift
    Leader, Plaza Senayan)**, **Budi Santoso (Agent, Parking Attendant)**.
  contact:
...
  license:
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /client-companies:
    get:
      operationId: listClientCompanies
      tags: [client-companies]
      summary: List client companies (also serves as ClientCompany picker)
      description: |
        Doubles as the ClientCompany picker per §18. Supports filters by
        status and whether a shift leader is already assigned (`has_leader`) —
        the unassigned-company picker is used by E3 leader-assignment.
      x-rbac:
        roles: [hr_admin, super_admin, shift_leader]
        scope: company_or_global
      x-design-screens: []  # AUDIT BLOCKER: F2.3 screens entirely missing
      parameters:
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Cursor' }
        - { $ref: '#/components/parameters/Sort' }
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: status
          schema:
            type: string
            enum: [ACTIVE, INACTIVE]
        - in: query
          name: has_leader
          description: |
            `true` = company has an assigned shift leader (1 leader / company
            invariant from E3). `false` = unassigned (leader-picker use).
          schema: { type: boolean }
      responses:
        '200':
          description: Page of client companies.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PageEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/ClientCompany' }
              examples:
                two_companies:
                  value:
                    data:
                      - id: SWP-CMP-101
                        name: Plaza Group
                        address: "Jl. M.H. Thamrin Kav. 28-30, Jakarta Pusat 10350"
                        leader_scope: site
                        npwp: "01.234.567.8-013.000"
                        pic_name: Ahmad Rizki
                        phone: "+62-21-572-5555"
                        email: ops@plazagroup.example.com
                        status: ACTIVE
                        has_leader: true
                        site_count: 3
                        active_placement_count: 42
                      - id: SWP-CMP-102
                        name: Mall Kelapa Gading
                        address: "Jl. Boulevard Raya, Jakarta Utara 14240"
                        leader_scope: company
                        npwp: null
                        pic_name: null
                        phone: null
                        email: null
                        status: ACTIVE
                        has_leader: false
                        site_count: 1
                        active_placement_count: 0
                    next_cursor: null
                    has_more: false
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      operationId: createClientCompany
      tags: [client-companies]
      summary: Create a client company (CC-1)
      description: |
        Creates a client company and **auto-provisions a primary "Main Site"**
        (F2.6 / CC-1c) so it is immediately placeable. Geofence config lives on
        Sites — set it via `POST /client-companies/{id}/sites` or by editing the
        Main Site. `leader_scope` defaults to `company`.
      x-rbac:
        roles: [hr_admin, super_admin]
        scope: global
      x-design-screens: []  # AUDIT BLOCKER
      parameters:
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ClientCompanyWriteRequest' }
            examples:
              plaza_group:
                value:
                  name: Plaza Group
                  address: "Jl. M.H. Thamrin Kav. 28-30, Jakarta Pusat 10350"
                  leader_scope: site
                  npwp: "01.234.567.8-013.000"
                  pic_name: Ahmad Rizki
                  phone: "+62-21-572-5555"
                  email: ops@plazagroup.example.com
      responses:
        '201':
          description: Created.
          headers:
            Location:
              schema: { type: string, format: uri }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ClientCompany' }
        '400':
          description: Bad request — e.g., geofence radius out of bounds.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
              examples:
                bad_radius:
                  value:
                    error:
                      code: GEOFENCE_RADIUS_INVALID
                      message: "Radius geofence harus 25–1000 meter."
                      fields:
                        geofence_radius_m: "Nilai 5000 di luar rentang yang diizinkan."
                      request_id: req_01J5XLA2
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: Duplicate name or NPWP (CC-2).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /client-companies/{client_company_id}:
    parameters:
      - { $ref: '#/components/parameters/ClientCompanyIdPath' }
    get:
      operationId: getClientCompany
      tags: [client-companies]
      summary: Get client company detail
      x-rbac:
        roles: [hr_admin, super_admin, shift_leader]
        scope: company_or_global
      responses:
        '200':
          description: Client company detail.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ClientCompany' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      operationId: updateClientCompany
      tags: [client-companies]
      summary: Edit client company (statutory fields, leader_scope)
      description: |
        Partial update — statutory/billing fields, PIC contact, and
        `leader_scope` (company ↔ site). Geofence/location is edited on the
        company's Sites (F2.6), not here.
      x-rbac:
        roles: [hr_admin, super_admin]
        scope: global
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ClientCompanyWriteRequest' }
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ClientCompany' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Duplicate name or NPWP.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /client-companies/{client_company_id}:deactivate:
    parameters:
      - { $ref: '#/components/parameters/ClientCompanyIdPath' }
    post:
      operationId: deactivateClientCompany
      tags: [client-companies]
      summary: Deactivate a client company (CC-5 — active-placement guard)
      description: |
        Sets the company to `INACTIVE`. **Blocks** if active placements exist
        unless `force=true` is passed AND the actor accepts the consequence
        (returns `COMPANY_HAS_ACTIVE_PLACEMENTS` (409) by default). HR is
        expected to end/transfer placements first per CC-5.
      x-rbac:
        roles: [hr_admin, super_admin]
        scope: global
      parameters:
        - { $ref: '#/components/parameters/IdempotencyKey' }
        - in: query
          name: force
          description: |
            If `true`, deactivate even with active placements (will mark them
            for review). Default `false` (warn-and-guide per CC-5 open
            decision).
          schema: { type: boolean, default: false }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string }
      responses:
        '200':
          description: Deactivated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ClientCompany' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Active placements present and `force` not set.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
              examples:
                active_placements:
                  value:
                    error:
                      code: COMPANY_HAS_ACTIVE_PLACEMENTS
                      message: "Plaza Senayan masih memiliki 42 penempatan aktif. Akhiri atau pindahkan terlebih dahulu."
                      fields:
                        active_placement_count: "42"
                      request_id: req_01J5XLB3
  /client-companies/{client_company_id}:reactivate:
    parameters:
      - { $ref: '#/components/parameters/ClientCompanyIdPath' }
    post:
      operationId: reactivateClientCompany
      tags: [client-companies]
      summary: Reactivate a client company (C-2)
      x-rbac:
        roles: [hr_admin, super_admin]
        scope: global
      parameters:
        - { $ref: '#/components/parameters/IdempotencyKey' }
      responses:
        '200':
          description: Reactivated — selectable as placement target again.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ClientCompany' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Already active.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }

  # =====================================================================
  # CLIENT SITES & GEOFENCE (F2.6)
  # =====================================================================
  /client-companies/{client_company_id}/sites:
    parameters:
      - { $ref: '#/components/parameters/ClientCompanyIdPath' }
    get:
      operationId: listSites
      tags: [sites]
      summary: List a company's sites (also serves as the Site picker)
      description: |
        Sites of a client company (F2.6). Doubles as the **Site picker** for
        placement (E3) per §18 — the primary "Main Site" is returned first.
      x-rbac:
        roles: [hr_admin, super_admin, shift_leader]
        scope: company_or_global
      parameters:
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Cursor' }
        - in: query
          name: status
          schema: { type: string, enum: [ACTIVE, INACTIVE] }
      responses:
        '200':
          description: Page of sites (primary first).
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PageEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Site' }
              examples:
                two_sites:
                  value:
                    data:
                      - id: SWP-SITE-501
                        client_company_id: SWP-CMP-101
                        name: Plaza Senayan
                        code: PS
                        address: "Jl. Asia Afrika No. 8, Jakarta Pusat 10270"
                        geo: { lat: -6.2253, lng: 106.7995 }
                        geofence_radius_m: 100
                        geofence_active: true
                        is_primary: true
                        pic_name: Ahmad Rizki
                        phone: "+62-21-572-5555"
                        status: ACTIVE
                        active_placement_count: 42
                      - id: SWP-SITE-502
                        client_company_id: SWP-CMP-101
                        name: Plaza Senayan Annex
                        code: null
                        address: "Jl. Asia Afrika No. 10, Jakarta Pusat 10270"
                        geo: null
                        geofence_radius_m: 100
                        geofence_active: false
                        is_primary: false
                        pic_name: null
                        phone: null
                        status: ACTIVE
                        active_placement_count: 0
                    next_cursor: null
                    has_more: false
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      operationId: createSite
      tags: [sites]
      summary: Add a site to a client company (ST-1)
      description: |
        Creates a site. `geofence_radius_m` defaults to 100m (§8 lock); valid
        range 25–1000m → else `GEOFENCE_RADIUS_INVALID` (400). `geofence_active`
        is derived from `geo` presence (ST-8). A site with no geo is allowed —
        geofencing is just disabled until coordinates are added. `name` is
        unique within the company (ST-2).
      x-rbac:
        roles: [hr_admin, super_admin]
        scope: global
      parameters:
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SiteWriteRequest' }
            examples:
              senayan:
                value:
                  name: Plaza Senayan
                  address: "Jl. Asia Afrika No. 8, Jakarta Pusat 10270"
                  geo: { lat: -6.2253, lng: 106.7995 }
                  geofence_radius_m: 100
                  pic_name: Ahmad Rizki
      responses:
        '201':
          description: Created.
          headers:
            Location: { schema: { type: string, format: uri } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Site' }
        '400':
          description: Bad request — e.g., geofence radius out of bounds.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
              examples:
                bad_radius:
                  value:
                    error:
                      code: GEOFENCE_RADIUS_INVALID
                      message: "Radius geofence harus 25–1000 meter."
                      fields: { geofence_radius_m: "Nilai 5000 di luar rentang yang diizinkan." }
                      request_id: req_01J5XLA2
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Duplicate site name within the company (ST-2).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /sites/{site_id}:
    parameters:
      - { $ref: '#/components/parameters/SiteIdPath' }
    get:
      operationId: getSite
      tags: [sites]
      summary: Get site detail
      x-rbac:
        roles: [hr_admin, super_admin, shift_leader]
        scope: company_or_global
      responses:
        '200':
          description: Site detail.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Site' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      operationId: updateSite
      tags: [sites]
      summary: Edit a site (geofence, location, contact)
      description: |
        Partial update — set `geo` to activate the geofence, adjust
        `geofence_radius_m`, update the on-site PIC. Set `is_primary: true` to
        make this the primary site (demotes the current primary atomically).
      x-rbac:
        roles: [hr_admin, super_admin]
        scope: global
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SiteWriteRequest' }
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Site' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Duplicate site name within the company (ST-2).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /sites/{site_id}:deactivate:
    parameters:
      - { $ref: '#/components/parameters/SiteIdPath' }
    post:
      operationId: deactivateSite
      tags: [sites]
      summary: Deactivate a site (ST-6 — active-placement + last-site guard)
      description: |
        Sets the site `INACTIVE`. **Blocks** (409) if active placements exist
        (`SITE_HAS_ACTIVE_PLACEMENTS`) or if it is the company's last active
        site while the company is active (`SITE_IS_LAST_ACTIVE`). The primary
        flag must be moved to another active site first.
      x-rbac:
        roles: [hr_admin, super_admin]
        scope: global
      parameters:
        - { $ref: '#/components/parameters/IdempotencyKey' }
      responses:
        '200':
          description: Deactivated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Site' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Active placements present, or this is the last active site.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
              examples:
                active_placements:
                  value:
                    error:
                      code: SITE_HAS_ACTIVE_PLACEMENTS
                      message: "Plaza Senayan masih memiliki 42 penempatan aktif. Akhiri atau pindahkan terlebih dahulu."
                      fields: { active_placement_count: "42" }
                      request_id: req_01J5XLB3

  # =====================================================================
  # POSITION TYPEAHEAD (free-text)
  # =====================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-or-jwt
      description: |
        `Authorization: Bearer <token>` per CONVENTIONS §3. Token type
        (opaque vs JWT) is deferred to E1; consumers MUST treat as opaque.
  parameters:
    Limit:
      in: query
      name: limit
      description: Page size (default 50, max 200) per §8.
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Cursor:
      in: query
      name: cursor
      description: Opaque cursor from a previous page's `next_cursor` per §8.
      schema: { type: string }
    Sort:
      in: query
      name: sort
      description: |
        Comma-separated `field:direction` per §9 (e.g.,
        `?sort=created_at:desc,full_name:asc`). Cursor encodes sort — mismatch
        returns `CURSOR_MISMATCH` (400).
      schema: { type: string }
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: false
      description: UUID v4. Same key with different body → `IDEMPOTENCY_KEY_REUSED` (409). Per §13.
      schema: { type: string, format: uuid }
...
  schemas:
...
    PageEnvelope:
      type: object
      description: Standard pagination envelope (CONVENTIONS §8).
      properties:
        data:
          type: array
          items: {}
        next_cursor:
          type: [string, "null"]
          description: Opaque cursor for the next page; `null` when no more pages.
        has_more:
          type: boolean
      required: [data, has_more]
...
    ClientCompany:
      type: object
      description: |
        Client company directory entry (F2.3). Holds statutory/billing info only —
        the **physical locations + geofence live on its Sites** (F2.6, relocated
        2026-06-03; see the `Site` schema and `/client-companies/{id}/sites`).
      required: [id, name, address, leader_scope, status]
      properties:
        id: { $ref: '#/components/schemas/ClientCompanyId' }
        name: { type: string, maxLength: 200, example: Plaza Group }
        address:
          type: string
          description: Registered / billing address (not the geofence — that's per Site).
          example: "Jl. M.H. Thamrin Kav. 28-30, Jakarta Pusat 10350"
        leader_scope:
          type: string
          enum: [company, site]
          default: company
          description: |
            Shift-leadership granularity (E3 F3.4): `company` = one leader for the
            whole company; `site` = one leader per active Site. Default `company`.
        npwp:
          type: [string, "null"]
          description: Optional Indonesian tax ID. Unique when set (CC-2).
        pic_name: { type: [string, "null"] }
        phone: { type: [string, "null"] }
        email: { type: [string, "null"], format: email }
        status:
          type: string
          enum: [ACTIVE, INACTIVE]
        has_leader:
          type: boolean
          readOnly: true
          description: True iff the company has an assigned shift leader (company-scope) (E3).
        site_count:
          type: integer
          readOnly: true
          description: Number of active sites (always ≥1 — every company has a primary Main Site, INV-5).
        active_placement_count:
          type: integer
          readOnly: true
          description: Cached count for the warn-and-guide deactivate flow (CC-5).
        created_at: { type: string, format: date-time, readOnly: true }
        updated_at: { type: string, format: date-time, readOnly: true }
...